# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library:
    get:
      operationId: getLibraryConfig
      tags: [library]
      summary: Get library configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Library config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryConfig"
    put:
      operationId: updateLibraryConfig
      tags: [library]
      summary: Update library configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryConfigInput"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryConfig"
  /public/libraries/{id}:
    get:
      operationId: getPublicLibrary
      tags: [library]
      summary: Get public library profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public library profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryConfig"
        "404":
          description: Library not found
...
  /seats:
    get:
      operationId: getSeatLayout
      tags: [seats]
      summary: Get all seats with occupancy
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [occupied, vacant, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Seat layout
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seat"
  /seats/config:
    put:
      operationId: updateSeatsConfig
      tags: [seats]
      summary: Set total number of seats
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatsConfigInput"
      responses:
        "200":
          description: Seats configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: recordPayment
      tags: [payments]
      summary: Record a new payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    LibraryConfig:
      type: object
      required: [libraryName, totalSeats, upiId, upiQrUrl, setupCompleted, language, theme]
      properties:
        libraryName:
          type: string
        totalSeats:
          type: integer
        upiId:
          type: ["string", "null"]
        upiQrUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        setupCompleted:
          type: boolean
        language:
          type: string
        theme:
          type: string
        ownerName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        openingTime:
          type: ["string", "null"]
        closingTime:
          type: ["string", "null"]
        monthlyFeeDefault:
          type: ["number", "null"]
        libraryType:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        facilities:
          type: array
          items:
            type: string
        googleMapsLink:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        logoPublicId:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        coverPublicId:
          type: ["string", "null"]
        galleryImages:
          type: array
          items:
            type: string
        websiteEnabled:
          type: boolean
        websiteSlug:
          type: ["string", "null"]
    LibraryConfigInput:
      type: object
      properties:
        libraryName:
          type: string
        totalSeats:
          type: integer
        upiId:
          type: string
        upiQrUrl:
          type: string
        whatsappNumber:
          type: string
        address:
          type: string
        phone:
          type: string
        setupCompleted:
          type: boolean
        language:
          type: string
        theme:
          type: string
        ownerName:
          type: string
        city:
          type: string
        state:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        monthlyFeeDefault:
          type: number
        libraryType:
          type: string
        description:
          type: string
        facilities:
          type: array
          items:
            type: string
        googleMapsLink:
          type: string
        logoUrl:
          type: string
        logoPublicId:
          type: string
        coverImageUrl:
          type: string
        coverPublicId:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        websiteEnabled:
          type: boolean
        websiteSlug:
          type: string
    Student:
      type: object
      required: [id, name, phone, fatherName, seatNumber, joiningDate, monthlyFee, feeDueDate, feeStatus]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        photoUrl:
          type: ["string", "null"]
        seatNumber:
          type: integer
        joiningDate:
          type: string
        monthlyFee:
          type: number
        feeDueDate:
          type: integer
        feeStatus:
          type: string
          enum: [paid, unpaid, overdue]
        nextDueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Seat:
      type: object
      required: [seatNumber, isOccupied]
      properties:
        seatNumber:
          type: integer
        isOccupied:
          type: boolean
        status:
          type: string
          enum: [vacant, paid, due-soon, overdue]
        student:
          $ref: "#/components/schemas/Student"
    SeatsConfigInput:
      type: object
      required: [totalSeats]
      properties:
        totalSeats:
          type: integer
    Payment:
      type: object
      required: [id, studentId, amount, paymentDate, receiptNumber, month, year]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: string
        seatNumber:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        receiptNumber:
          type: string
        month:
          type: integer
        year:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [studentId, amount, month, year]
      properties:
        studentId:
          type: string
        amount:
          type: number
        month:
          type: integer
        year:
          type: integer
        paymentDate:
          type: string
        notes:
          type: string
    PaymentListResponse:
      type: object
      required: [payments, total, page, limit]
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...